+++ CHANGELOG.md
@@ -1,8 +1,24 @@
 # Changelog
 
-## [v1.0.0](https://github.com/couchbasecloud/terraform-provider-couchbase-capella/tree/v1.0.0) (2024-04-02)
+## [v1.2.1](https://github.com/couchbasecloud/terraform-provider-couchbase-capella/tree/v1.2.1) (2024-05-07)
 
-[Full Changelog](https://github.com/couchbasecloud/terraform-provider-couchbase-capella/compare/v1.0.0...v1.0.0)
+[Full Changelog](https://github.com/couchbasecloud/terraform-provider-couchbase-capella/compare/v1.2.0...v1.2.1)
+
+**Fixed bugs:**
+
+- \[AV-77464\] fix export schema [\#191](https://github.com/couchbasecloud/terraform-provider-couchbase-capella/pull/191) ([l0n3star](https://github.com/l0n3star))
+
+## [v1.2.0](https://github.com/couchbasecloud/terraform-provider-couchbase-capella/tree/v1.2.0) (2024-04-25)
+
+[Full Changelog](https://github.com/couchbasecloud/terraform-provider-couchbase-capella/compare/v1.1.0...v1.2.0)
+
+**Implemented enhancements:**
+
+- \[AV-73782\] Couchbase Server Audit Events Support [\#158](https://github.com/couchbasecloud/terraform-provider-couchbase-capella/pull/158) ([l0n3star](https://github.com/l0n3star))
+
+## [v1.1.0](https://github.com/couchbasecloud/terraform-provider-couchbase-capella/tree/v1.1.0) (2024-04-02)
+
+[Full Changelog](https://github.com/couchbasecloud/terraform-provider-couchbase-capella/compare/v1.0.0...v1.1.0)
 
 **Implemented enhancements:**